Makefile fixes: 0.85.1 recut was unusable.
When transitioning to git, I failed to test the build from a fresh
tarball extraction, so did not notice that the new SCRIPTSRC vs
SCRIPTDIST distinction made install impossible, because requisite source
files for regenerating a target were not available.

Simplest fix is also the most portable: strip out all targets intended
for distributors.  The segregated sections of the makefile have paid
off, it's a simple sed command to strip out.  So renamed "Makefile" to
"makefile" for the git repo, but the distributed version is a Makefile
with the distributor targets all ripped out, so that make(1) doesn't
think it has to rebuild the sources we distribute.

Additionally: when using install(1) to install to a directory, finish
the target path with "/./" so that if .../bin/ does not exist we don't
install the script as a file named "bin".  "/." is sufficient for my
install(1), but probably most portable to ensure target ends with a / if
we want directory semantics for how install(1) decides what to do.