SSL certs location: bugfixes, check more places.
Issue #1 is a report from LEVAI Daniel that the recent certs-finding
functionality is broken on OpenBSD, because of an assumption that if
we can find OPENSSLDIR from openssl(1), then certs/ within that must
be valid.

Made a validity check for a given dir; not existing is silent, but
existing and unusable triggers debug messages.  It checks for being
a dir and having filenames ending ".0" inside that directory.

Bug-fixed stderr-handling around the search.

If openssl doesn't give us a location, and none of the file candidates
exist, then check a few locations for SSL_ca_path directory candidacy.

Fixes #1