Be clearer about port specification meaning
The debug messages understandably led a user astray, because when seeing
`sieve(4190)` they thought that meant port 4190, not that this was the
port specification being passed down ("try to find 'sieve' in
/etc/services (or other getservent source) and use 4190 as a fallback").

Fix the debug message to notice when the port is being given in this
form and tack on an explanation.  Issue a debug warning, the first time
evaluating this, if services lists a different value, to highlight the
conflict and aid in debugging.

Fix the debug message for SRV records to not use the previous port
definition in the 'old' status logging, because the port number did not
factor into the SRV lookup and the claim was slightly off.

Thanks to Daniel Schulte for his patience when using the tool to debug.