Remote sieve name should be basename of local
*POSSIBLY BREAKING CHANGE*

If user specifies a path to a local sieve script, then we were using the
full path for upload to the remote sieve server.  Now we use the
basename, so everything after the last directory separator.

RFC 5804 does not disallow a directory separator, but timsieved does.
It seems likely that most servers will.

A remote name with directory separators can still be explicitly
specified.  This change tries to make the inferred name choice smarter.

Folks using `/` or `\` in path-names, where the remote server allows
this, may be affected.  In practice, I believe that breakage is
unlikely.