Require TLS verification to work
Philipp Erbelding filed a bug-report and PR to fix that if TLS
verification was requested but no CA file or path is available,
sieve-connect should die.  Which was good, but not enough.

Our handling here was to silently fall back to non-TLS cleartext if the
server did not advertise the STARTTLS capability.  That was an unhappy
compromise back when many did not configure TLS and the difficulty of
persuading that TLS should be configured was high(er).  But the world
has moved on, so let's try to fix this.  Hey, our major number is still
0.

With this:
 * unless `--clearchan` is passed on the command-line, if TLS
   verification is enabled then `--noclearchan` is implicitly assumed
   + unless `$SIEVECONNECT_INSECURE_CLEARTEXT_FALLBACK` set non-empty
     in environment
   + so, can change invocation, or set an obnoxious educational
     environment variable, but the default is better
 * Failure to find CA information means verification must fail, so abort
   before even trying to connect (unless `--clearchan` enabled per
   above)
 * slightly better debug logging for when this bites people

Tested via temporarily commenting out `tls_cert_file` in `imapd.conf`
for Cyrus IMAP.