Exit 1 if warnings seen if non-interactive
If used interactively, humans make silly typos and recover.  If not
interactive, those problems are actual errors which should cause us to
exit non-zero for tooling to be able to rely upon `$?`.

(I'm embarrassed to have let this slip through, but it's a result of the
 history: initially interactive, then scriptability added.)

To avoid missing some corner case, we just capture all calls to warn()
if non-interactive, and track things that way, instead of introducing a
new function.  Ah Perl, ♡ you so.