Release hacks via env: ignore copyright year, etc
Last change, for 0.89, was small enough that I wasn't comfortable
bumping the copyright year for it, so didn't, but that broke release
because of the sensible sanity-check.

So be able to override that check, and be able to readily pretend that
the repo is clean, since I had to override repo-generate to do so.